+++ test/test_reads.py
@@ -930,8 +930,408 @@ class TestFind(TestCase):
         start offset into a gapped sequence and 'end' is true.
         """
         self.assertEqual(
-            Read("id", "AC-ac--A-A--C-GT").find("AC", ignoreGaps=True, start=5, end=True),
+            Read("id", "AC-ac--A-A--C-GT").find(
+                "AC", ignoreGaps=True, start=5, end=True
+            ),
             13,
+        )
+
+
+class TestReadFindPrefixAndSuffix(TestCase):
+    """
+    Test the getPrefixAndSuffixOffsets method of the Read (note: singular) class.
+    """
+
+    def testNotMatched(self):
+        """
+        getPrefixAndSuffixOffsets must return -1, -1 if the prefix and suffix don't
+        match.
+        """
+        self.assertEqual(
+            Read("id", "ACGT").getPrefixAndSuffixOffsets("XXX", "YYY"), (-1, -1)
+        )
+
+    def testOnlyPrefixMatched(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offset and -1 if only the
+        prefix matches.
+        """
+        self.assertEqual(
+            Read("id", "ACTTGT").getPrefixAndSuffixOffsets("CTT", "YYY"), (1, -1)
+        )
+
+    def testOnlySuffixMatched(self):
+        """
+        getPrefixAndSuffixOffsets must return -1 and the correct offset if only the
+        suffix matches.
+        """
+        self.assertEqual(
+            Read("id", "ACTTGT").getPrefixAndSuffixOffsets("YYY", "CTT"), (-1, 4)
+        )
+
+    def testPrefixAndSuffixMatched(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offsets if both the prefix
+        and suffix match.
+        """
+        self.assertEqual(
+            Read("id", "ACTTGT").getPrefixAndSuffixOffsets("CT", "TG"), (1, 5)
+        )
+
+    def testNotMatchedIgnoringGaps(self):
+        """
+        getPrefixAndSuffixOffsets must return -1, -1 if the prefix and suffix don't
+        match and we are ignoring gaps.
+        """
+        self.assertEqual(
+            Read("id", "A-C--GT").getPrefixAndSuffixOffsets(
+                "XXX", "YYY", ignoreGaps=True
+            ),
+            (-1, -1),
+        )
+
+    def testOnlyPrefixMatchedIgnoringGaps(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offset and -1 if only the
+        prefix matches and gaps are ignored.
+        """
+        self.assertEqual(
+            Read("id", "AC-T--TGT").getPrefixAndSuffixOffsets(
+                "CTT", "YYY", ignoreGaps=True
+            ),
+            (1, -1),
+        )
+
+    def testOnlySuffixMatchedIgnoringGaps(self):
+        """
+        getPrefixAndSuffixOffsets must return -1 and the correct offset if only the
+        suffix matches.
+        """
+        self.assertEqual(
+            Read("id", "A--C-T-TGT").getPrefixAndSuffixOffsets(
+                "YYY", "CTT", ignoreGaps=True
+            ),
+            (-1, 8),
+        )
+
+    def testPrefixAndSuffixMatchedIgnoringGaps(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offsets if both the prefix
+        and suffix match.
+        """
+        self.assertEqual(
+            Read("id", "A-C-T-T--GT").getPrefixAndSuffixOffsets(
+                "CT", "TG", ignoreGaps=True
+            ),
+            (2, 10),
+        )
+
+
+class TestReadsFindPrefixAndSuffix(TestCase):
+    """
+    Test the getPrefixAndSuffixOffsets and extractRegion methods of the Reads
+    (note: plural) class.
+    """
+
+    def makeReads(self, *sequences) -> ReadsInRAM:
+        return ReadsInRAM(
+            Read(f"id-{i}", sequence) for i, sequence in enumerate(sequences)
+        )
+
+    def testNoReads(self):
+        """
+        getPrefixAndSuffixOffsets must raise a ValueError if the Reads instance
+        has no reads.
+        """
+        error = r"^No input sequences were given\.$"
+        with self.assertRaisesRegex(ValueError, error):
+            Reads().getPrefixAndSuffixOffsets("XXX", "YYY")
+
+    def testPrefixConflict(self):
+        """
+        getPrefixAndSuffixOffsets must raise a ValueError if the prefix is found
+        at different offsets.
+        """
+        reads = self.makeReads(
+            "ACGT",
+            "AAAC",
+        )
+
+        error = (
+            "^Conflict: prefix 'AC' was found at offset 0 in 'id-0' but at offset 2 in "
+            r"'id-1'\.$"
+        )
+        with self.assertRaisesRegex(ValueError, error):
+            reads.getPrefixAndSuffixOffsets("AC", "YYY")
+
+    def testSuffixConflict(self):
+        """
+        getPrefixAndSuffixOffsets must raise a ValueError if the suffix is found
+        at different offsets.
+        """
+        reads = self.makeReads(
+            "ACGT",
+            "AAAC",
+        )
+
+        error = (
+            "^Conflict: suffix 'AC' was found ending at offset 2 in 'id-0' but ending "
+            r"at offset 4 in 'id-1'\.$"
+        )
+        with self.assertRaisesRegex(ValueError, error):
+            reads.getPrefixAndSuffixOffsets("YYY", "AC")
+
+    def testOnlyPrefixMatched(self):
+        """
+        getPrefixAndSuffixOffsets must raise a ValueError if only the prefix
+        is matched.
+        """
+        reads = self.makeReads(
+            "ACGT",
+            "ACAT",
+        )
+
+        self.assertEqual(
+            reads.getPrefixAndSuffixOffsets("AC", "YYY"),
+            (
+                (0, -1),
+                [
+                    ((0, -1), reads[0]),
+                    ((0, -1), reads[1]),
+                ],
+            ),
+        )
+
+    def testOnlySuffixMatched(self):
+        """
+        getPrefixAndSuffixOffsets must raise a ValueError if only the suffix
+        is matched.
+        """
+        reads = self.makeReads(
+            "ACGT",
+            "ACAT",
+        )
+
+        self.assertEqual(
+            reads.getPrefixAndSuffixOffsets("YYY", "T"),
+            (
+                (-1, 4),
+                [
+                    ((-1, 4), reads[0]),
+                    ((-1, 4), reads[1]),
+                ],
+            ),
+        )
+
+    def testPrefixAndSuffixMatched(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offsets if both the prefix
+        and suffix match.
+        """
+        reads = self.makeReads(
+            "ACAGT",
+            "ACACT",
+            "ACACA",
+            "XXXXX",
+        )
+        self.assertEqual(
+            reads.getPrefixAndSuffixOffsets("CA", "T"),
+            (
+                (1, 5),
+                [
+                    ((1, 5), reads[0]),
+                    ((1, 5), reads[1]),
+                    ((1, -1), reads[2]),
+                    ((-1, -1), reads[3]),
+                ],
+            ),
+        )
+
+    def testSequenceIdNotFound(self):
+        """
+        getPrefixAndSuffixOffsets must raise a ValueError if the prefix and suffix
+        are supposed to be in a specific sequence but there is no sequence with the
+        specified id.
+        """
+        reads = self.makeReads(
+            "ACGT",
+            "ACAT",
+        )
+
+        error = r"^No sequence with id 'dummy-id' found\.$"
+        with self.assertRaisesRegex(ValueError, error):
+            reads.getPrefixAndSuffixOffsetsForId("dummy-id", "YYY", "T")
+
+    def testPrefixAndSuffixMatchedForSpecificId(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offsets if both the prefix
+        and suffix match in the specified read.
+        """
+        reads = self.makeReads(
+            "ACAGT",
+            "ACACT",
+        )
+        self.assertEqual(
+            reads.getPrefixAndSuffixOffsetsForId(reads[0].id, "CA", "T"),
+            ((1, 5), reads[0]),
+        )
+
+    def testPrefixAndSuffixMatchedForSpecificIdInconsistentOffsets(self):
+        """
+        getPrefixAndSuffixOffsets must return the correct offsets if both the prefix
+        and suffix match in the specified read, even if the prefix and suffix exist
+        in other sequences at different offsets (i.e., matching in the passed id is
+        all that matters).
+        """
+        reads = self.makeReads(
+            "ACAGT",
+            "CACTG",
+        )
+        self.assertEqual(
+            reads.getPrefixAndSuffixOffsetsForId(reads[0].id, "CA", "T"),
+            ((1, 5), reads[0]),
+        )
+
+    def testExtractRegionUnevenLengths(self):
+        """
+        extractRegion must raise a ValueError if the sequences are of uneven length.
+        """
+        sequences = (
+            "ACAGT",
+            "ACACTA",
+            "ACACA",
+            "XXXXX",
+        )
+
+        reads = self.makeReads(*sequences)
+        error = (
+            "^All sequences must be the same length, unless "
+            r"allowUnequalLengths is true. Found lengths 5, 6\.$"
+        )
+        with self.assertRaisesRegex(ReadLengthsNotIdenticalError, error):
+            reads.extractRegion(id_=None, prefix="CA", suffix=None)
+
+    def testExtractRegionUnevenLengthsAllowed(self):
+        """
+        extractRegion must return the correctly trimmed reads if only the prefix
+        matches and the sequences are of uneven length but that is indicated as
+        a non-error.
+        """
+        sequences = (
+            "ACAGT",
+            "ACACTA",
+            "ACACA",
+            "XXXXX",
+        )
+
+        reads = self.makeReads(*sequences)
+        expectedReads = self.makeReads(*[s[1:] for s in sequences])
+        result, offsets, details = reads.extractRegion(
+            id_=None, prefix="CA", suffix=None, allowUnequalLengths=True
+        )
+
+        self.assertEqual(offsets, (1, -1))
+
+        self.assertTrue(all(r1 == r2 for (r1, r2) in zip(result, expectedReads)))
+
+        self.assertEqual(
+            details,
+            [
+                ((1, -1), reads[0]),
+                ((1, -1), reads[1]),
+                ((1, -1), reads[2]),
+                ((-1, -1), reads[3]),
+            ],
+        )
+
+    def testExtractRegionOnlyPrefixMatched(self):
+        """
+        extractRegion must return the correctly trimmed reads if only the prefix
+        matches.
+        """
+        sequences = (
+            "ACAGT",
+            "ACACT",
+            "ACACA",
+            "XXXXX",
+        )
+
+        reads = self.makeReads(*sequences)
+        expectedReads = self.makeReads(*[s[1:] for s in sequences])
+        result, offsets, details = reads.extractRegion(id_=None, prefix="CA", suffix=None)
+
+        self.assertTrue(all(r1 == r2 for (r1, r2) in zip(result, expectedReads)))
+
+        self.assertEqual(offsets, (1, -1))
+
+        self.assertEqual(
+            details,
+            [
+                ((1, -1), reads[0]),
+                ((1, -1), reads[1]),
+                ((1, -1), reads[2]),
+                ((-1, -1), reads[3]),
+            ],
+        )
+
+    def testExtractRegionOnlySuffixMatched(self):
+        """
+        extractRegion must return the correctly trimmed reads if only the suffix
+        matches.
+        """
+        sequences = (
+            "ACAGT",
+            "ACACT",
+            "ACACA",
+            "XXXXX",
+        )
+
+        reads = self.makeReads(*sequences)
+        expectedReads = self.makeReads(*[s[:5] for s in sequences])
+        result, offsets, details = reads.extractRegion(id_=None, prefix=None, suffix="T")
+
+        self.assertEqual(offsets, (-1, 5))
+
+        self.assertTrue(all(r1 == r2 for (r1, r2) in zip(result, expectedReads)))
+
+        self.assertEqual(
+            details,
+            [
+                ((-1, 5), reads[0]),
+                ((-1, 5), reads[1]),
+                ((-1, -1), reads[2]),
+                ((-1, -1), reads[3]),
+            ],
+        )
+
+    def testExtractRegionPrefixAndSuffixMatched(self):
+        """
+        extractRegion must return the correctly trimmed reads if both the prefix
+        and suffix match.
+        """
+        sequences = (
+            "ACAGT",
+            "ACACT",
+            "ACACA",
+            "XXXXX",
+        )
+
+        reads = self.makeReads(*sequences)
+        expectedReads = self.makeReads(*[s[1:5] for s in sequences])
+        result, offsets, details = reads.extractRegion(id_=None, prefix="CA", suffix="T")
+
+        self.assertTrue(all(r1 == r2 for (r1, r2) in zip(result, expectedReads)))
+
+        self.assertEqual(offsets, (1, 5))
+
+        self.assertEqual(
+            details,
+            [
+                ((1, 5), reads[0]),
+                ((1, 5), reads[1]),
+                ((1, -1), reads[2]),
+                ((-1, -1), reads[3]),
+            ],
         )
 
 
@@ -2813,9 +3213,9 @@ class TestReads(TestCase):
         read2 = Read("id2", "AC")
         initialReads = Reads([read1, read2])
         initialReads.filter(minLength=3)
+        reads = Reads(initialReads)
 
         read3 = Read("id3", "AC")
-        reads = Reads(initialReads)
         reads.add(read3)
         self.assertEqual(sorted((read1, read3)), sorted(reads))
         self.assertEqual(3, reads.unfilteredLength())